refactor: Move method from balance helpers to group helpers
Helps with circular dependency since we will need getGroupBalance
in translateAndSortGroups
