+++ src/ducks/balance/helpers.spec.js
@@ -5,7 +5,6 @@ import {
   sumBalanceHistories,
   getAllDates,
   balanceHistoryToChartData,
-  getGroupBalance,
   getPanelsState
 } from './helpers'
 import { format as formatDate, parse as parseDate } from 'date-fns'
@@ -271,43 +270,6 @@ describe('balanceHistoryToChartData', () => {
     const values = chartData.map(item => item.y)
 
     expect(values).toEqual(expected)
-  })
-})
-
-describe('getGroupBalance', () => {
-  it('should return 0 if no group is given', () => {
-    expect(getGroupBalance()).toBe(0)
-  })
-
-  it('should return 0 if the group has no account', () => {
-    expect(getGroupBalance({ accounts: null })).toBe(0)
-    expect(getGroupBalance({ accounts: { data: null } })).toBe(0)
-  })
-
-  it('should return the sum of all accounts balance', () => {
-    const accounts = [{ balance: 1000 }, { balance: -100 }, { balance: 8000 }]
-    const group = { accounts: { data: accounts } }
-
-    expect(getGroupBalance(group)).toBe(8900)
-  })
-
-  it('should not sum the excluded accounts balance', () => {
-    const accounts = [
-      { _id: 'a', balance: 1000 },
-      { _id: 'b', balance: -100 },
-      { _id: 'c', balance: 8000 }
-    ]
-
-    const group = { accounts: { data: accounts } }
-    const excludedAccounts = ['a', 'c']
-
-    expect(getGroupBalance(group, excludedAccounts)).toBe(-100)
-  })
-
-  it('should return 0 if the group has no account', () => {
-    const group = { accounts: { data: [] } }
-
-    expect(getGroupBalance(group)).toBe(0)
   })
 })
